compliance with master

+++ src/joybuttontypes/joygradientbutton.cpp
@@ -25,6 +25,7 @@
 
 #include <QDebug>
 
+
 JoyGradientButton::JoyGradientButton(int index, int originset, SetJoystick *parentSet, QObject *parent) :
     JoyButton(index, originset, parentSet, parent)
 {
@@ -62,6 +63,7 @@ void JoyGradientButton::turboEvent()
             if (isKeyPressed)
             {
                 changeState = false;
+
                 if (!turboTimer.isActive() || (turboTimer.interval() != 5))
                 {
                     turboTimer.start(5);
@@ -88,25 +90,26 @@ void JoyGradientButton::turboEvent()
             {
                 if (getTurboMode() == GradientTurbo)
                 {
-                    tempInterval2 = static_cast<int>(floor((getMouseDistanceFromDeadZone() * turboInterval) + 0.5));
+                    tempInterval2 = floor((getMouseDistanceFromDeadZone() * turboInterval) + 0.5);
                 }
                 else
                 {
-                    tempInterval2 = static_cast<int>(floor((turboInterval * 0.5) + 0.5));
+                    tempInterval2 = floor((turboInterval * 0.5) + 0.5);
                 }
             }
             else
             {
                 if (getTurboMode() == GradientTurbo)
                 {
-                    tempInterval2 = static_cast<int>(floor(((1 - getMouseDistanceFromDeadZone()) * turboInterval) + 0.5));
+                    tempInterval2 = floor(((1 - getMouseDistanceFromDeadZone()) * turboInterval) + 0.5);
                 }
                 else
                 {
                     double distance = getMouseDistanceFromDeadZone();
+
                     if (distance > 0.0)
                     {
-                        tempInterval2 = static_cast<int>(floor(((turboInterval / getMouseDistanceFromDeadZone()) * 0.5) + 0.5));
+                        tempInterval2 = floor(((turboInterval / getMouseDistanceFromDeadZone()) * 0.5) + 0.5);
                     }
                     else
                     {
@@ -121,6 +124,7 @@ void JoyGradientButton::turboEvent()
                 // remaining time left.
                 tempTurboInterval = tempInterval2 - turboHold.elapsed();
                 int timerInterval = qMin(tempTurboInterval, 5);
+
                 if (!turboTimer.isActive() || (turboTimer.interval() != timerInterval))
                 {
                     turboTimer.start(timerInterval);
@@ -165,11 +169,11 @@ void JoyGradientButton::turboEvent()
                 {
                     if (getTurboMode() == GradientTurbo)
                     {
-                        tempTurboInterval = static_cast<int>(floor((getMouseDistanceFromDeadZone() * turboInterval) + 0.5));
+                        tempTurboInterval = floor((getMouseDistanceFromDeadZone() * turboInterval) + 0.5);
                     }
                     else
                     {
-                        tempTurboInterval = static_cast<int>(floor((turboInterval * 0.5) + 0.5));
+                        tempTurboInterval = floor((turboInterval * 0.5) + 0.5);
                     }
 
                     int timerInterval = qMin(tempTurboInterval, 5);
@@ -198,18 +202,20 @@ void JoyGradientButton::turboEvent()
                 releaseDeskEvent();
 
                 isKeyPressed = false;
+
                 if (turboTimer.isActive())
                 {
                     if (getTurboMode() == GradientTurbo)
                     {
-                        tempTurboInterval = static_cast<int>(floor(((1 - getMouseDistanceFromDeadZone()) * turboInterval) + 0.5));
+                        tempTurboInterval = floor(((1 - getMouseDistanceFromDeadZone()) * turboInterval) + 0.5);
                     }
                     else
                     {
                         double distance = getMouseDistanceFromDeadZone();
+
                         if (distance > 0.0)
                         {
-                            tempTurboInterval = static_cast<int>(floor(((turboInterval / getMouseDistanceFromDeadZone()) * 0.5) + 0.5));
+                            tempTurboInterval = floor(((turboInterval / getMouseDistanceFromDeadZone()) * 0.5) + 0.5);
                         }
                         else
                         {
@@ -282,9 +288,10 @@ void JoyGradientButton::wheelEventVertical()
         else if ((diff >= 0.1) && (wheelSpeedY != 0))
         {
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedY / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedY / distance;
             }
             else
             {
@@ -297,6 +304,7 @@ void JoyGradientButton::wheelEventVertical()
                 // remaining time left.
                 tempInterval = tempInterval - wheelVerticalTime.elapsed();
                 tempInterval = qMin(tempInterval, 5);
+
                 if (!mouseWheelVerticalEventTimer.isActive() || (mouseWheelVerticalEventTimer.interval() != tempInterval))
                 {
                     mouseWheelVerticalEventTimer.start(tempInterval);
@@ -313,15 +321,17 @@ void JoyGradientButton::wheelEventVertical()
     if ((buttonslot != nullptr) && (wheelSpeedY != 0))
     {
         bool isActive = getActiveSlots().contains(buttonslot);
+
         if (isActive && activateEvent)
         {
             sendevent(buttonslot, true);
             sendevent(buttonslot, false);
             mouseWheelVerticalEventQueue.enqueue(buttonslot);
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedY / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedY / distance;
             }
             else
             {
@@ -329,6 +339,7 @@ void JoyGradientButton::wheelEventVertical()
             }
 
             tempInterval = qMin(tempInterval, 5);
+
             if (!mouseWheelVerticalEventTimer.isActive() || (mouseWheelVerticalEventTimer.interval() != tempInterval))
             {
                 mouseWheelVerticalEventTimer.start(tempInterval);
@@ -342,10 +353,12 @@ void JoyGradientButton::wheelEventVertical()
     else if (!mouseWheelVerticalEventQueue.isEmpty() && (wheelSpeedY != 0))
     {
         QQueue<JoyButtonSlot*> tempQueue;
+
         while (!mouseWheelVerticalEventQueue.isEmpty())
         {
             buttonslot = mouseWheelVerticalEventQueue.dequeue();
             bool isActive = getActiveSlots().contains(buttonslot);
+
             if (isActive && activateEvent)
             {
                 sendevent(buttonslot, true);
@@ -362,9 +375,10 @@ void JoyGradientButton::wheelEventVertical()
         {
             mouseWheelVerticalEventQueue = tempQueue;
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedY / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedY / distance;
             }
             else
             {
@@ -372,6 +386,7 @@ void JoyGradientButton::wheelEventVertical()
             }
 
             tempInterval = qMin(tempInterval, 5);
+
             if (!mouseWheelVerticalEventTimer.isActive() || (mouseWheelVerticalEventTimer.interval() != tempInterval))
             {
                 mouseWheelVerticalEventTimer.start(tempInterval);
@@ -436,9 +451,10 @@ void JoyGradientButton::wheelEventHorizontal()
         else if ((diff >= 0.1) && (wheelSpeedX != 0))
         {
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedX / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedX / distance;
             }
             else
             {
@@ -451,6 +467,7 @@ void JoyGradientButton::wheelEventHorizontal()
                 // remaining time left.
                 tempInterval = tempInterval - wheelHorizontalTime.elapsed();
                 tempInterval = qMin(tempInterval, 5);
+
                 if (!mouseWheelHorizontalEventTimer.isActive() || (mouseWheelHorizontalEventTimer.interval() != tempInterval))
                 {
                     mouseWheelHorizontalEventTimer.start(tempInterval);
@@ -473,9 +490,10 @@ void JoyGradientButton::wheelEventHorizontal()
             sendevent(buttonslot, false);
             mouseWheelHorizontalEventQueue.enqueue(buttonslot);
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedX / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedX / distance;
             }
             else
             {
@@ -501,6 +519,7 @@ void JoyGradientButton::wheelEventHorizontal()
         {
             buttonslot = mouseWheelHorizontalEventQueue.dequeue();
             bool isActive = getActiveSlots().contains(buttonslot);
+
             if (isActive)
             {
                 sendevent(buttonslot, true);
@@ -513,9 +532,10 @@ void JoyGradientButton::wheelEventHorizontal()
         {
             mouseWheelHorizontalEventQueue = tempQueue;
             double distance = getMouseDistanceFromDeadZone();
+
             if (distance > 0.0)
             {
-                tempInterval = 1000 / wheelSpeedX / static_cast<int>(distance);
+                tempInterval = 1000 / wheelSpeedX / distance;
             }
             else
             {
